Fix vsync time/period narrow cast
AIDL definition of DisplayStatInfo has 'long' type for vsyncTime and
vsyncPeriod, and according to the aidl_to_cpp.cpp:GetCppName
'long' equals 'int64_t'. 'nsecs_t' is typedef for 'int64_t' too.

Hence, on platforms where C++ 'long' is 32bit we have narrow
conversion and lose data, while there is no need for cast at all.

+++ services/surfaceflinger/SurfaceFlinger.cpp
@@ -8915,8 +8915,8 @@ binder::Status SurfaceComposerAIDL::getDisplayStats(const sp<IBinder>& display,
     DisplayStatInfo statInfo;
     status_t status = mFlinger->getDisplayStats(display, &statInfo);
     if (status == NO_ERROR) {
-        outStatInfo->vsyncTime = static_cast<long>(statInfo.vsyncTime);
-        outStatInfo->vsyncPeriod = static_cast<long>(statInfo.vsyncPeriod);
+        outStatInfo->vsyncTime = statInfo.vsyncTime;
+        outStatInfo->vsyncPeriod = statInfo.vsyncPeriod;
     }
     return binderStatusFromStatusT(status);
 }
